Moved scatter series into a subfolder.
