+++ js/Series/MapSeries.js
@@ -23,7 +23,7 @@ import SVGRenderer from '../Core/Renderer/SVG/SVGRenderer.js';
 import U from '../Core/Utilities.js';
 var extend = U.extend, fireEvent = U.fireEvent, getNestedProperty = U.getNestedProperty, isArray = U.isArray, isNumber = U.isNumber, merge = U.merge, objectEach = U.objectEach, pick = U.pick, splat = U.splat;
 import '../Core/Options.js';
-import './ScatterSeries.js';
+import './Scatter/ScatterSeries.js';
 /**
  * @private
  * @class
