+++ js/Series/PolygonSeries.js
@@ -16,7 +16,7 @@ import LegendSymbolMixin from '../Mixins/LegendSymbol.js';
 import LineSeries from './Line/LineSeries.js';
 import '../Core/Options.js';
 import '../Core/Legend.js';
-import './ScatterSeries.js';
+import './Scatter/ScatterSeries.js';
 /**
  * A polygon series can be used to draw any freeform shape in the cartesian
  * coordinate system. A fill is applied with the `color` option, and
