+++ ts/Series/Scatter3DSeries.ts
@@ -51,7 +51,7 @@ declare module '../Core/Series/SeriesType' {
     }
 }
 
-import './ScatterSeries.js';
+import './Scatter/ScatterSeries.js';
 
 /**
  * @private
