+++ ts/Series/TreemapSeries.ts
@@ -362,7 +362,7 @@ declare module '../Core/Series/SeriesType' {
 }
 
 import '../Core/Options.js';
-import './ScatterSeries.js';
+import './Scatter/ScatterSeries.js';
 
 /* eslint-disable no-invalid-this */
 const AXIS_MAX = 100;
